fix(02C): address re-Codex round 2 — cost-cast crash, anti-mezcla, idempotent fan-out, audit, loop-cap
P1 cost_usd::float8 (pg numeric→string crashed the Escalated list's toFixed in prod). P1-4 anti-mezcla (§3.5): runMotorForCohort rejects a stale cohort_rule_version (motor.run takes an arbitrary cohortId; fn_nba_test_all reads current-version signals). P2-2: a cohort/action dedup CONFLICT is an idempotent 'skipped', not a bogus 'dispatch_failed' that floods the human queue (new outcome + skipped tally). P2-1: Security_Log on the blocked cross-pool cohort (§3.4 audit). P1-6: motor_max_loops bound capped 1..3 (the contract/UI/message promise ≤3). KEEP (Leo): confidence-threshold knob + 'resolved'-on-dispatch (rationale + measurement-producer follow-up documented). Defer: P1-2/P1-3 multi-version policy semantics (latent in single-version demo).

+++ server/motor/controls.ts
@@ -81,7 +81,7 @@ export async function listEscalations(tenantId: string, exec: Exec): Promise<Mot
     `select kc.kb_case_id::text as kb_case_id, kc.area_type, coalesce(kc.pattern,'') as pattern,
             kc.not_resolved_reason, kc.discarded_branches, c.cost as cost_usd, kc.created_at::text as created_at
        from tenant."Knowledge_Case" kc
-       left join (select ref_id, sum(cost_usd) as cost from gov.v_llm_cost group by ref_id) c
+       left join (select ref_id, sum(cost_usd)::float8 as cost from gov.v_llm_cost group by ref_id) c
               on c.ref_id = kc.path_used->>'attempt_id'
       where kc.tenant_id=$1 and kc.outcome='escalated'
       order by kc.created_at desc limit 50`,
